requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.3.1
anyio==4.11.0
asgiref==3.10.0
attrs==23.2.0
billiard==4.2.3
CacheControl==0.14.4
cachetools==5.5.2
celery==5.3.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
cryptography==46.0.3
dj-database-url==3.0.1
Django==4.2.23
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
django-leaflet==0.32.0
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.14.0
djangorestframework-gis==1.2.0
fcm-django==2.0.0
feedparser==6.0.10
firebase-admin==6.4.0
frozenlist==1.4.1
geographiclib==2.1
geopy==2.4.1
google-api-core==2.28.1
google-api-python-client==2.128.0
google-auth==2.29.0
google-auth-httplib2==0.2.1
google-cloud-core==2.5.0
google-cloud-firestore==2.14.0
google-cloud-storage==2.16.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.62.3
gunicorn==21.2.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
joblib==1.5.2
kombu==5.6.0
msgpack==1.1.2
multidict==6.0.5
nltk==3.8.1
packaging==25.0
pillow==10.3.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyfcm==2.0.9
PyJWT==2.10.1
pyparsing==3.2.5
python-crontab==3.3.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2025.2
redis==5.0.1
regex==2025.11.3
requests==2.31.0
rsa==4.9.1
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.66.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.6.0
yarl==1.9.4
